Se refactoriza inventory page y se divide servidor de cliente, actualmente esta separado pero hay que platicar si desea que se separe aun mas
